added marker point to selected vertex started mode selection function begrudgingly removed casting to single - can cause unintended side effects...

+++ matSurf_fns/@brainSurf/calcCentroid.m
@@ -9,7 +9,7 @@ function calcCentroid(obj,vert)
 vert = [vert;mean(vert)];
 
 % split the volume into tetrahedrons (pyramids)
-T = uint32(delaunay(vert));
+T = delaunay(vert);
 
 % calculate centroids of each tetrahedron
 centroids = [...
